requirements.txt
asgiref==3.11.0
Django==5.0.6
django-tinymce==5.0.0
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.4
tzdata==2025.3
whitenoise==6.11.0
